# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "gobudget API",
        "title": "gobudget API",
        "contact": {},
        "version": "beta"
    },
    "host": "localhost:8080",
    "basePath": "/beta",
    "paths": {
...
        "/budgets/{budget_id}/transactions": {
            "get": {
                "description": "List all transactions across all accounts in the budget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "List all transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "budget_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.TransactionsView"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Create a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "budget_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction details",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TransactionsView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/budgets/{budget_id}/transactions/{transaction_id}": {
            "get": {
                "description": "Get the details of a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "budget_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "TransactionRequest": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "category_id",
                "date",
                "payee_id"
            ],
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "string"
                },
                "cleared": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "payee_id": {
                    "type": "string"
                },
                "reconciled": {
                    "type": "boolean"
                }
            }
        },
        "TransactionResponse": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "approved": {
                    "type": "boolean"
                },
                "category_name": {
                    "type": "string"
                },
                "cleared": {
                    "type": "boolean"
                },
                "date": {
                    "$ref": "#/definitions/pgtype.Date"
                },
                "memo": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "payee_name": {
                    "type": "string"
                },
                "reconciled": {
                    "type": "boolean"
                }
            }
        },
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "db.TransactionsView": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "account_name": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "approved": {
                    "type": "boolean"
                },
                "budget_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "cleared": {
                    "type": "boolean"
                },
                "date": {
                    "$ref": "#/definitions/pgtype.Date"
                },
                "id": {
                    "type": "string"
                },
                "memo": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "payee_id": {
                    "type": "string"
                },
                "payee_name": {
                    "type": "string"
                },
                "reconciled": {
                    "type": "boolean"
                }
            }
        },
...
        "pgtype.Date": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.InfinityModifier": {
            "type": "integer",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Text": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}